ferray-fft: split plan cache locks and accept negative axes
- Split each plan cache from a single Mutex<{planner, plans}> into a
  read-mostly RwLock<HashMap<key, Arc<plan>>> plus a Mutex<Planner>
  taken only on the rare cache-miss path. The previous design serialized
  every concurrent FFT lookup even when the plan was already cached,
  which produced contention under heavy multi-threaded workloads
  (#431). Both the complex (rustfft) and real (realfft) caches use
  the same pattern. Type aliases keep clippy quiet.
- Switch the public `axis: Option<usize>` / `axes: Option<&[usize]>`
  parameters on every FFT entry point (fft, ifft, fft2, ifft2, fftn,
  ifftn, rfft, irfft, rfft2, irfft2, rfftn, irfftn, hfft, ihfft,
  fftshift, ifftshift, fft_real*) to `Option<isize>` /
  `Option<&[isize]>`. Negative values are normalized in
  `axes::normalize_axis` so `axis=-1` selects the last axis exactly
  as NumPy does (#434). This is a breaking change to the public API,
  but the crate is on 0.2.x and the prior signature was a NumPy
  compatibility gap that needed fixing.
- Two new unit tests verify `fft(..., Some(-1))` and
  `fftn(..., Some(&[-2, -1]))` produce results identical to the
  positive-axis equivalents.

Closes #431, #434.

+++ ferray-fft/src/hermitian.rs
@@ -44,7 +44,7 @@ use crate::norm::FftNorm;
 pub fn hfft<T: FftFloat, D: Dimension>(
     a: &Array<Complex<T>, D>,
     n: Option<usize>,
-    axis: Option<usize>,
+    axis: Option<isize>,
     norm: FftNorm,
 ) -> FerrayResult<Array<T, IxDyn>>
 where
@@ -78,7 +78,9 @@ where
         FftNorm::Ortho => FftNorm::Ortho,
     };
 
-    crate::real::irfft::<T, IxDyn>(&conj_arr, Some(output_len), Some(ax), hfft_norm)
+    // `ax` is already validated against `ndim` here, so casting to
+    // `isize` is lossless within usize::MAX/2.
+    crate::real::irfft::<T, IxDyn>(&conj_arr, Some(output_len), Some(ax as isize), hfft_norm)
 }
 
 /// Compute the inverse FFT of a real-valued signal, returning
@@ -100,7 +102,7 @@ where
 pub fn ihfft<T: FftFloat, D: Dimension>(
     a: &Array<T, D>,
     n: Option<usize>,
-    axis: Option<usize>,
+    axis: Option<isize>,
     norm: FftNorm,
 ) -> FerrayResult<Array<Complex<T>, IxDyn>>
 where
@@ -120,7 +122,7 @@ where
         FftNorm::Ortho => FftNorm::Ortho,
     };
 
-    let result = crate::real::rfft::<T, D>(a, n, Some(ax), ihfft_norm)?;
+    let result = crate::real::rfft::<T, D>(a, n, Some(ax as isize), ihfft_norm)?;
 
     // Conjugate the output
     let conj_data: Vec<Complex<T>> = result.iter().map(|c| c.conj()).collect();
